pkg/osbuild: move mount name generation into getOsbuildMount
Mount names are just identifiers. They don't have any other real
meaning.

Prior this commit, callers of genOsbuildMount had to supply the mount
name, which led to inconsistencies: genMountsDevicesFromPt used the
name of the leaf device, genMountsForBootupd used synthetic names
constructed from a partition number.

The issue with the genMountsDevicesFromPt's approch is that multiple
mounts can be created from a simple device. This is not currently
used, but it will come with btrfs (single device) and its subvolumes
(multiple mounts). The genMountsForBootupd's approach has a similar
flaw.

This commit implements the mount name generation directly in
genOsbuildMount. The new name is no longer referring to the source,
but it's rather constructed from the mountpoint (by using pathEscape).
This has two benefits:
- Every mount obviously must have a mountpoint
- Mountpoints are unique(*), so names will be also unique

I also added some docstrings, because these functions signatures are
quite hard to comprehend.

(*) Yeah, you can stack mounts, but this is not something supported
by our disk library.