cmdutil: simplify SeedArgFor to take a string instead of a `namer`
Simplify the previous commit so that SeedArgFor() takes strings
for the image type and distribution instead of the `namer`
interface. This makes the code a little bit easier to follow but
now the `distro.Distro` and `distro.ImageType` cannot be directly
passed into the helper anymore so it's a tradeoff.

Ideally it would just take `distro.{Distro,ImageType}` but for
that we would need an easy way to construct fake objects of
these types in our tests which there (currently) is not (AFAICT).

So settle with this approach as a compromise for now.