fedora: add dbus-broker to iot-simplified-installer dracut modules
With the sources update, the iot-simplified-installer in Fedora 40
started failing in the dracut stage with:

    dracut[I]: Module 'dbus-daemon' will not be installed, because command 'dbus-daemon' could not be found!
    dracut[E]: Module 'dbus' depends on 'dbus-daemon', which can't be installed
    dracut[E]: Module 'network-manager' depends on 'dbus', which can't be installed
    ...

This seems to be related to this change (which is included in dracut 060
in Fedora 40):
https://github.com/dracutdevs/dracut/commit/3f8f115a27b2bb63f0893262b6c951a187bac8ef

The dbus dracut module has a dependency on either dbus-daemon or
dbus-broker.  In the past, simply installing dbus-broker was enough to
satisfy the dependency because the `find_binary` function would find the
dbus-broker binary.  The change to `check_module` requires the module to
be explicitly enabled, otherwise it falls back to dbus-daemon.