manifest: signal end of bootc install on serial port
Out automatic tests need a way to know when the install is finished.
Before we could just trigger on a qmp:REBOOT event but with this
change the installer will wait for user input before doing a reboot.
Ejecting the cdrom tray is also done after the user confirmation
so triggering on qmp:DEVICE_TRAY_MOVED is also not an option.

Go the conservative route and just output on the serial port. This
has the added benefit that users on the serial port also get an
indication when the install is done.