test/generate-ostree-build-config: error on unknown distro name
For ostree-based systems we need a default ref that depends on  the
product name (iot or edge).  So far, if it wasn't set, we would get an
error (`product` would be undefined).  Let's explicitly raise the error.
If a distro ever gets added that includes an ostree build and is not one
of the three existing ones, we should know to update this.