Set the chunk size for S3 upload to 64MB
The default minimum chunk size of 5MB for s3 upload only allows for an
upload of about 50Gi. It is not out of the question for a bootc image to
get above this threshhold.

For now, this will increase the chunk size to 64MB making it possible to
upload images up to ~640Gi. This could in the future become a dynamic
value, but for now, this should at least give some head room.

Signed-off-by: Brad P. Crochet <brad@redhat.com>