otk-gen-partition-table: add "modifications"
This commit adds a new `modifications` field to the input of the
`otk-gen-partition-table` that can be used to tweak the output of the
partition generator.

The inputs follow the inputs from the existing customizations, so
the json input looks like this:
```json
{
  ...
  "modifications": {
    "partition_mode": "raw",
    "filesystems": [
      {"mountpoint": "/var/log", "minsize": 10241024}
    ]
  }
}
```

and the otk snippet would look something like:
```yaml
filesystem:
  otk.external.gen_partition_table:
    options: ...
    partitions: ...
    modifications:
      partition_mode: ${modifications.partition_mode}
      filesystems: ${modfications.filesystems}
```
and the user would override them via:
```
cat > mymod.json <<'EOF'
{
  "modifications": {
    "partition_mode": "raw",
    "filesystems: [ {"mountpoint": "/var/log", "minsize": 10241024 } ]
}
$ otk compile --modifications=mymod.json centos-9-x86_64-ami.otk
```

Also, renames InputOptions ("options") to InputProperties ("properties")
to clarify that these fields are not configuration options that can be
be used to modify partition tables but instead should be properties tied
to the base partition table itself and changes to these should be
handled through the modifications.