New command: otk-gen-partition-table
New Partition Table generator command that makes the internal generation
code available to external tooling (otk).

The command takes as input a base partition table (like the ones defined
in pkg/distro/**/partition-tables.go) and a set of options and
customizations.  These match the arguments to the
disk.NewPartitionTable() function.

The output is a fully resolved Partition Table (with all UUIDs, start
and offset values, and mount options defined), a set of kernel options,
and a partition map.

The kernel options are useful for some stages that require adding kernel
arguments that include the root partition UUID.
The partition map is a convenient mapping of names to partitions with
predictable names (root and boot) so that otk and other externals can
access properties specific to those partitions.

The fully resolved Partition Table itself is nested under the "internal"
key.  This is meant to signify that the representation of the full
Partition Table is an internal data structure that shouldn't be
considered stable or usable by otk itself.  Instead, it is meant to be
fed back into other commands that will use it to generate useful data
structures (like stage options, mounts, and devices).

Co-authored-by: Achilleas Koutsou <achilleas@koutsou.net>