test/generate-ostree-build-config: download oci archive from s3
Download the container archive from the s3 bucket instead of pulling it
from the gitlab registry.  After this, we will phase out use of the
gitlab registry and start removing all the containers we have there.

Each archive is downloaded to a temporary directory and run directly
with `podman run oci-archive:<path>`, which will also pull the archive
into the registry, then the temporary directory (along with the
downloaded archive) is discarded.