disk: support json unmarshal for `Partition`
The `disk.Partition` uses interfaces inside the data structure.
To support json unmarshal we need a custom `UnmarshalJSON()`
that peeks into the marshalled json and detects what concrete
type to use.

Go makes this slightly complicated, this commit uses a
list of all known `disk.Entity` types to detect what is
supported and reflect to try to build the data. It's not
super efficient as it will try all known disk.Entities
but the advantage is that the only requirement to add
new entities is that they are added to the known entities
slice.

This commit provides the `disk.Partition.UnmarshalJSON`
implementation and a test.