image: allow unset kickstart options where possible
For the AnacondaContainerInstaller, having empty kickstart options when
creating the ImageKind should be possible.  The only option that's
required is the Kickstart.Path, which has a default value we can set.

For the AnacondaOSTreeInstaller, at least the OSTree options are
required, so return an error when they're not set, but also set the
Kickstart.Path to the default when it's not already set from the caller.