manifest: add OTKManifest type
The OTKManifest is a second implementation of the manifest.Manifest
interface that will be used to generate manifests from otk yaml files.
The type only implements the Serialize() method.  None of the other
methods are important (at least for now).  The content spec methods in
particular are not needed since otk will handle content resolution
(depsolving, container and ostree commit resolves) itself.