bootupd: Don't require /boot mount point
I'd like to change bib to just drop the separate `/boot` partition
by default because it isn't doing anything useful right now.

A separate `/boot` is only useful with things like LUKS or LVM,
but since bib doesn't do those now and it's not in the near term
roadmap clearly either, let's just not add the complexity
of a separate `/boot` by default.

Signed-off-by: Colin Walters <walters@verbum.org>