image: assert that serialization panics when props aren't set
The tests assert that the serialization of installer image kinds panics
when the ISO-related properties aren't set (product name, product
version, and ISO label).  The panics happen in the stage validation, but
we might want to catch them earlier (perhaps make them mandatory in the
image kind or pipeline constructors) in the future.

For now, we just want to ensure that invalid manifests (with empty
strings) aren't generated.