pkg: add otkdistro
New package that loads a distribution from otk yaml files.
Distribution-level properties, like name, codename, versions, etc, are
loaded from a file called 'properties.yaml' in the root of the
distribution's directory.  An error is returned if this file is missing
or if any of the required properties are undefined/empty.

The distro creates a distro->arch->imagetype hierarchy based on the
contents of the filesystem tree.  It assumes a directory structure
`<distroid>/<archname>/<imagetypename>.yaml`.  For example:

fedora-40/
  properties.yaml
  x86_64/
    qcow2.yaml
    ami.yaml
  aarch64/
    qcow2.yaml

The top-level directory is used as the unique distro ID, whereas the
name is defined in the properties.yaml.  The distinction is useful for
separating the on-disk, unique ID of a distro from the display name.