pkg: add otk for handling otk-related things
New otk package for managing otk-related things.
The otkCustomizationsFile type handles creating an otk file with defines
for customizations and a single otk.include.  The purpose is to use this
to create a 'customizations.yaml' file that 'otk.define's all the
variables represented by a blueprint then 'otk.include' the file we want
to compile, effectively setting variables defined by the user in the
compilation of the manifest.

Currently it only writes a few simple customizations as a proof of
concept, but it will grow to cover all potential customizations.