image: add tests for Anaconda installer images
Test that setting the AdditionalanacondaModules and
DisabledAnacondaModules properties on the three installer image kinds
results in the expected stage options.  This replicates the test cases
in manifest for the AnacondaInstaller pipeline.  The purpose is to
ensure that the image properties are correctly "wired up" to the
pipeline and the subsequent stage creation.