disk: add Btrfs to the types implementing VolumeContainer
Implement the VolumeContainer interface for Btrfs.  Btrfs doesn't have a
fixed metadata size that needs to be reserved, like the LUKS header.
Similar to the LVM VolumeGroup, there is a lot that can affect the size
of the metadata in a btrfs volume.  However, there is no fixed location
or size that we need to reserve.  More importantly, btrfs subvolumes
don't need to partition the space of a btrfs volume, which means we
don't need to reserve metadata space on the partition to fit metadata
alongside the required size of the sum of subvolumes, like we need to do
for LVM.