pkg/blueprints: remove xml tailoring
Following a meeting with compliance and the OpenSCAP team we decided to
implement the json tailoring. They are trying to hide and abstract away
the xml tailoring completely for future versions of RHEL and Fedora. So
we decided to remove the xml tailoring support - this can easily be
added back in if we need it at a later stage