disk: fix package docstring
Conventionally, the first line of a go package's docstring starts with

  Package <package name>

and is a short summary of the package that is shown in the list of all
packages.

There should be no space between the docstring block and the package
name declaration, otherwise the comment block is ignored
(see for example the docs in the current latest tag:
https://pkg.go.dev/github.com/osbuild/images@v0.73.0/pkg/disk)

This commit also rephrases a couple of sentences and fixes a typo.