Move the logic handling RHSMConfig from OS pipeline to distro
Historically, the RHSMConfig is defined for two scenarios in the default
image definition - when the image should be subscribed and when not. The
config for both scenarios has been previously copied to the
`OSCustomizations` struct and the appropriate RHSMConfig would be picked
when serializing the `OS` pipeline.

Simplify the flow and make `OSCustomizations` struct contain only the
infal RHSMConfig that should be applied to the image, if any. Move the
logic deciding which RHSMConfig should be used, to the function handling
the creation of `OSCustomizations` from the BP customizations, default
`ImageConfig` and `ImageOptions`. This means to the `osCustomizations()`
function, which is part of the distro implementation.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>