rhel10: drop hardcoded UUIDs from {centos,rhel}-10
Historically the rhel/centos image defintion hardcoded the partition
table, partition and filesystem UUIDs. The exact reasons for this
is unclear and it seems it was done mostly for comaptibility with
the previous generation of image build tools.

However given that a UUIDs is meant to be (practically) unique
from rhel10 on we stop hardcoding them. The images library will
then create a random one for each image build. For the cases
where this is undesirable the `OSBUILD_TESTING_RNG_SEED` env
can be used to control the RNG.