fedora: drop hardcoded UUIDs for partitions/filesystems
Similar to the change in the previous commit for centos10/rhel10
the hardcoding of the UUIDs is also dropped on fedora.

Ideally we would only do it for fedora-41 or later but the
library is not split this way yet and just introducing a new
factory for fc41 seems overkill (but I'm happy to do it if
others feel it's worth it).