disk: make `GetFSTabOptions()` return an error
The disk package is now used for the `otk` externals. This means
that the inputs we get are much less controlled and instead of
panicing we should return more user-friendly error messages.

This commit starts by making `GetFSTabOptions()` return an error
and then bubbling it up.

Followup to https://github.com/osbuild/images/pull/787/files#r1673752935