otk-make-fstab-stage: remove `recover()`
With the changes in the previous commit to return an error from
`osbuild.NewFSTabStageOptions` if something goes wrong there is
no need anymore for the `recover()` that was used before.

Note that it is currently not entirely easy to reason about this
because one needs to manually follow all possible code paths
into `osbuild` and `disk` to see what functions could `panic()`.
For this one it was realtivly small, for the `otk-make-partition-stages`
it is a lot more involved though.