internal: add new `types.Option` generic for primitive types
This is heavily inspired by https://github.com/moznion/go-optional
with is super nice. However we cannot use it as is because it
lacks support for toml unmarshal and go makes it really hard to
extend existing types. Submiting toml unmarshal upstream would
also be an option but it's difficult in the general case because
the toml unmarshal interface is very limited for complex types
(it only passes `data any` with the full decoded complex struct
into map[string]inteface{} for example which leaves it to the
user of the library to convert to the actual type struct which
is a pain with reflect). I really whish go was more expressive
here.

So this commit defines a subset of https://github.com/moznion/go-optional
in our own internal types lib.