distro/rhel{9,10}: remove bogus rd.blacklist=nouveau kernel arg
rd.blacklist is actually not processed by anything. The correct option
is called rd.driver.blacklist, see
https://man.archlinux.org/man/dracut.8.en

Hovewer, we can actually safely remove it altogether, because we
also blacklist nouveau in a /etc/modprobe.d drop-in config. This is
actually the recommended way, see
https://access.redhat.com/solutions/41278

Thus, this commit removes the redundant, and actually bogus argument.