osbuild: only mount the required mountpoints for bootc install
bootc install to-filesystem requires a clean root to install, with only
the required mountpoints present: /, /boot, and /boot/efi.  Do not mount
any user-defined mountpoints when running the bootc install stage.

Content in bootc images cannot be separated across filesystems,
therefore there cannot be any content under the custom mountpoints, so
this change will not cause any issues of content not being on the
correct filesystem.