osbuild: add test for essential mounts in NewBootcInstallToFilesystemStage
Small tweak for the tests for `NewBootcInstallToFilesystemStage`
so that we also test that only the essential mounts are used in
the `org.osbuild.bootc.install-to-filesystem`. This also makes
the fake creation for /boot/efi more realistic by moving the fs
to `vfat`.

Tiny followup for https://github.com/osbuild/images/pull/846