manifest: New pipeline: subscription
The subscription pipeline is a very simple pipeline that creates files
(services, keyfile) needed to subscribe a system on boot.  Usually these
files are created in an OS tree when the subscription options are used.
The purpose of the new pipeline is to create the files in a clean tree
so that other pipelines can copy them to other locations.

The reason for doing this now is to copy the files onto an ISO so that
the files can be copied into the OS tree using a kickstart file %post
section to create the files at installation time when deploying an
ostree-based system.  This avoids making the files (which include
activation keys) part of the immutable base system so that they can be
removed after activation without needing to update the base system.

The file creation ownership is set to nil:nil (instead of root:root) to
avoid running an org.osbuild.chown stage in the subscription pipeline,
since the stage requires chroot and the pipeline is empty.