cmd: rename image(s) flag to type(s)
The build command requires the --image flag to specify the image type to
build.  Similarly, gen-manifests and list-images commands support
specifying image types using the --images flag.  These were originally
named this way because image-type(s) felt too long.  Rename them to
--type and --types.  This is less ambiguous and matches the flag name
chosen by bootc-image-builder.

The cmds are development and testing tools, so there is no need to keep
the old and deprecate it for a backwards compatible and smooth
transition.