manifest: add a README to the /subscription directory on the ISO
Add a README file under /subscription on the ISO that explains what the
files in the directory are meant for.  When copying the files out in the
kickstart %post, copy only the /subscription/etc/ directory contents so
the README doesn't get added to the system.

Also makes sure filenode stages are only created once in the
makeKickstartStages() function by collecting all files (on p.Files) and
generating the stages at the end of the function.