disk: test that mountpoints from the bp were applied
Previously, the test suite didn't check whether the customizations
were actually applied to a partition table. This commit adds a tiny
function that tests that the correct Muntable is created in the
partition table and that all Sizeables above the Mountable have at least
the MinSize given by the blueprint.