pkg: add 'sbom' package for working with SBOM documents
Add a new `sbom` package for working with SBOM documents. It provides a
very simple wrapper struct, which currently supports only SPDX standard.
The SBOM document is for now stored in a raw JSON form, to not have to
convert the raw data from and to the specific in-memory representation
on the worker.

The idea is to provide a bit of an abstraction from the specific SBOM
implementation, so that in the future, it would be possible to create
`sbom.Document` from SBOM documents of various standards and also
serialize it back to various SBOM standards.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>