disk: fallback to higher-level containers when creating new filesystems
Prior this commit, when the first-found mountpoint creator (a container
nearest to the root mountpoint) failed to create a new mountpoint for
some reason, the whole partition table customizing failed.

This commit changes this behaviour: if a mountpoint creation fails, the
code continues to traverse the path in the direction of the partition
table root, and tries to find another creator and use that one instead.

This is now mostly a pointless change because the only error can be
returned from an LVM volume group that fails to allocate a unique name
which is basically impossible. However, the following commits will use
this feature to correctly allocate swap.