{blueprint,disk}: add support for swap partitions
This commit allows the filesystem customization in a blueprint to add a
swap partition. This is done by adding a new field "type". It currently
has only two values: unset and swap.

The createFilesystem now creates a swap partition if type == swap.
Otherwise, it still falls back to hardcoded "xfs". You can clearly see
that should not be hard to change. ;)

An interesting case is when / is on btrfs. Then, the library will try to
create swap as a btrfs subvolume, which is bogus, so it returns an
error. However, one of the previous commits changed the logic, so a
higher-level container is tried in this case. So swap will be created as
a raw partition as usual (or LV, if you have btrfs on LV).

The blueprintApplied test helper can now work with swap, and it was
also added into one of the test blueprints, so all the new code should
be covered.