blueprint: make json filesystem parsing consistent with the toml one
There's no functional change except for a minor adjustements of an
error. I just made sure that everything refers to minsize instead of
size. Also, I removed a comment that mentioned that the keys are called
differently in json than in toml, but that's no longer true. See e.g.
the previous commit.