os: disable dracut for ostree based systems on rpm installation
Installing the `minimal-environment` RPM group implies the installation
of `dracut-config-rescue` which makes `org.osbuild.ostree.preptree` to
fail because `rpm-ostree` detects multiple kernels within the tree
(the usual one and the rescue one). Disabling dracut during the rpm
installation prevents the rescue kernel/initrd to be generated and
makes `rpm-ostree` to finish the compose succesfully.

`rpm-ostree` runs dracut on his own during the compose so this doesn't
seem to affect the final result.

Resolves: #624