cmd/otk-*: do basic input validation to avoid crashing too easily
This commit is a lesson from https://github.com/osbuild/otk/pull/186
as it is right now extremly easy to crash the otk disk externals
with bad inputs.

This commit adds basic input validation, it's probably a good
idea to also go deeper into the library and make it less trusting
about pointers but that is (slightly) orthogonal.