test: test edge-ami with ostree commit with @minimal-environment
The @minimal-environment rpm group includes 'dracut-config-rescue'
which, by default, will create a rescue kernel configuration.  With
ostree-based images, this will fail, since rpm-ostree requires that only
one kernel exists in an image.  We now disable dracut when installing
packages for ostree commits, so this should no longer be a problem.

Adding a configuration pair to test that the image builds and boots.
The configs include:
1. An edge-container built with the @minimal-environment rpm group,
   which will pull in the dracut-config-rescue package.  This is built
   for all RHEL 9 versions.
2. An edge-ami built using the new edge container as input.  The config
   includes a user so that the image can be booted in CI.