test/generate-ostree-build-config: fix podman stop command
The podman stop command was passing the list of container IDs as a
single argument with spaces in it, which is wrong.  The arguments need
to be separate, so we should expand the list of IDs instead.

This was never an issue because we never had multiple containers in a
distro's ostree build config until now.