anacond: move SElinux policy up
The previous commit introduced SElinux to the live installer. This
commit brings the pipeline more in line with the os pipeline, SElinux is
now a property on the pipeline set in the image.

It is hardcoded to be set to targeted for live installers and is
considered a programming error to be set for non-live installers.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>