cmd/build: add new `-manifest-only` option
This option allows build to just generate the manifest but not
run `osbuild`. This is useful for the `otk` tests where a manifest
for a specific distro needs to get generated to produce identical
otk files.

Note that we also have `gen-manifests` but that seems geared towards
a different use-case so it's harder to use/change for this purpose.