cmd/build: add `-target-arch` flag
This flag allows to set the target architecture for the build. This
is most useful for `-manifest-only` and the intended use-case is
generating reference images for `otk`. But it could also be used
with `qemu-user` and should work equally well as `bib` (maybe
better as the "regular" distro code tends to use less of the
modern system calls that tend to be problematic like `openat2`).