reporegistry: embed the test repo in the go binary for easier use
Running `cmd/build/build` currently requires running it from a git
checkout so that it can find the required repository files.

This commit embeds them into the binary via "go:embed" to make
it easier to use `build` on it's own.

This will make working with `otk` to generate reference images
easier.