manifest: remove explicit ownership from insights files
When creating the service drop-in file and directory for running the
insights client on boot, don't specify ownership.  This is for the same
reasons as for the other files we create in the subscription pipeline:
there is no `chown` binary in the subscription pipeline tree, which
would be required to explicitly set the file's ownership.

Update tests to match.