github: don't fail fast on Fedora unit tests
Fedora 41 unit tests are currently failing due to an issue with
the repositories and are non-essential for merging (which is common for
rawhide and early branching periods).
However, the fail-fast property defaults to true, which causes the
other Fedora tests to get cancelled when F41 fails.

Disable fail-fast so that F40 tests keep running when F41 tests fail.

This is the same change we made for CS tests
3b81a908327c740b95a86b630d993dfad31fe5cb