github: don't fail fast on cs unit tests
CentOS Stream 10 unit tests are currently failing due to an issue with
the repositories and we are making it non-essential for merging.
However, the fail-fast property defaults to true, which causes the
CentOS Stream 9 tests to get cancelled when 10 fails.

Disable fail-fast so that C9S tests keep running when C10S tests fail.