image: add new `gce` pipeline to `BootcDiskImage`
There is a request that we provide a new image type that can
be imported directly into GCE. This requires a `tar.gz` with
a `disk.raw` inside and specific tar options.

This commit adds the needed pipeline to generate this.

Note that this will require a way to rename files during
the tar build as we do not want to hardcode the raw image
filename.