Fix loading of tested distro repos for unit tests
The NewTestedDefault() was returning an empty repositories list when
used within unit tests. And unfortunately, this didn't trigger any error
and no test was checking that the list is not empty, in addition to
checking the error.

Modify `Reporegistry.New()` to return error if no repos were loaded.

Modify the `NewTestedDefault()` to consistently determine the path to
load repositories from, regardless of the situation in which it is being
called.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>