blueprint: remove the sshkey customization
Both the sshkey and users customizations have been around since the
beginning of time [1].  The users customization was always a superset of
the sshkey, so we have code in blueprints to convert sshkeys into users.

The blueprint code in images is not exposed directly to users.  The code
responsible for deserializing blueprint toml files lives in
osbuild-composer and is directly translated to the equivalent types in
osbuild/images.  We can simplify the code and types here to only contain
the customizations that are strictly necessary and we don't need to
maintain backwards compatibility.

The blueprint conversion code in osbuild-composer should convert all
sshkeys to users before passing them down to osbuild/images functions.

[1] 967b0e8ce0e87c77a7c681b0c9f86d70f84711a5