ostree: fix error check for rhsm.LoadSystemSubscriptions() call
The condition would panic if an error was returned from the function
because 'subs' would be nil.  Also, the underlying error would be lost
if it wasn't nil.

Handle 'err != nil' separately to avoid dereferencing a nil 'subs' and
append the underlying error.

Keep the 'subs.Consume' nil check as well since the function can succeed
but not attach Consumer secrets when they're not available.  'subs'
should never be nil when the function succeeds.