cmd/otk-resolve-ostree-commit: add more test variants
Added tests for calling the resolver with a commit ID (see below) with
and without a URL as well as error tests.

Resolving by ID: When a commit ID is used in place of the ref, the
resolver returns the input commit ID as both the ref and the checksum.
No request goes to the server for this call.