reporegistry: add NoReposLoadedError and use it
Add `NoReposLoadedError` custom error type for the case when any
function that is supposed to load repository configs, does not load
any. Use it in all functions that are directly loading repositories
from configuration files.

I considered `NoReposError` name, but that could be later added for
functions such as `RepoRegistry.ReposBy...()`, in case there are no
repositories in the RepoRegistry for the given image type / arch.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>