github: drop `-v` from `go test`
This commit removes `-v` from `go test`. The test output is very
long (17k lines according to GH) and most is just:
```
...
--- PASS: TestRPMDeduplication (0.00s)
...
```
but when trying to find a failing test (like in PR#945) scrolling
through to find the actual failing one becomes a bit cumbersome.

So this commit drops `-v`.