Makefile: initial version of makefile, replace GH action
This introduces the initial version of `Makefile` and
implements `make test` to be able to run the tests locally

also adds `make gh-action-test` to test-github actions
locally (in a container).

To accomplish this, a new Containerfile now replaces the steps
of the github action and the github action now just calls `make
gh-action-test`

This way we only have _one_ implementation of the
environment/dependency setup.