test: build edge-commit with parent
We never had a test configuration where an ostree commit is built as
child of another one.  Let's add a simple configuration where an ostree
commit is built from an existing one.

The parent commit is the base image with no customizations (empty) and
the child commit adds a user (edge-ostree-pull-user).