github: new workflow "Schutzbot comments"
The purpose of this new workflow is to run when certain actions are
allowed to cause errors without failing but PR authors and reviewers
should be notified of issues.  The goal is, instead of having actions
that are allowed to fail and are marked as not required, the actions
themselves will succeed but will communicate an error message through a
shared artifact that this action can read and use to post a comment on
the PR.

The action is run as a 'workflow_run' to have access to secrets
(Schutzbot's access token).

The action is not finalised in this commit.