github: save osbuild-composer integration test result as artifact
Save the status of the integration (reverse dependency) test with
osbuild-composer to an artifact along with the PR number.  This data
will be then used by the comment action to post a comment on the PR
informing the author and reviewers of the failure.  The job itself wont
fail if the unit tests fail.