blueprint: split Blueprint json/toml keys into their own structs
This commit separates the `bp.Blueprint` struct from the actual
json/toml keys associated with it. This will allow us to deprecate
keys and rename keys without having to change the bp.Blueprint
structure in the code. There is also only a single parsing function
that will be used by the callers that can ensure compatibility
and consistency.