test: pull tests from main and trim to RHEL 9.4
Updating the test configs and test scripts is necessary to be compatible
with recent changes in the way we handle testing ostree-based images.
Since a58470c85817734daedc559a2eb871ba0907c509 (v0.65.0), we stopped
using the gitlab registry to host ostree containers.

This commit also removes all repository configurations except RHEL 9.4
and regenerates the gitlab test matrix so that we only focus on testing
that distribution version.

The unit tests on Fedora 40 have been removed.  Unit tests will only run
on CentOS Stream 9 with an osbuild version matching the version used in
v0.40.0.

The config-map has been reduced to only include configurations valid for
RHEL 9.4.

The cmd/boot-aws command is also pulled from main because it will be
called with new flags and features from the test scripts in ./test.