distrosort: add new `distrosort.Names` helper
This commit adds a new `distrosort.Names()` helper that can be
used sort sort distro names. It will take version numbers of
the distro into account and sort by semantic version. This
is useful in e.g. user-interfaces where we want to present
the version in a nicely sorted way, e.g. a version like
`1.4` is sorted before `1.10` (which a `string.Sort()` gets
wrong).

Sorting follows the go-version library which supports semantic
versioning [0] so anything supported there will be sorted
correctly including `-beta` suffixes and similar (see the
spec for details). However this does not work today because
the underlying distroid parser does only support major/minor.
But once this is fixed full sort will work.

Note that this is an extra pkg because the `distro` and `distroid`
would be circular imports otherwise. The alternative might be to
create a distro.ID package.

[0] https://semver.org/