blueprint: split PartitioningCustomization validation function
Split the part of the function that validates that we only have one
btrfs subvolume and one LVM Volume Group.  The general Validate()
function now allows multiple VGs and btrfs volumes.

The reason for having two functions is to be able to have less strict
validation in certain contexts.  In the NewCustomPartitionTable()
function, which is a reusable function in the library, we use the
Validate() function which only checks for functionally invalid
configurations (bad mountpoint names or invalid structures).  Whether
multiple volumes are allowed is a policy decision that should be left up
to the distro definition.