disk: generate unique subvol names and labels for root
When adding a root partition or btrfs subvolume automatically, collect
all the labels and names in the partition table and generate a unique
name or label for the entity.

The new genUniqueString() function follows the logic used when
generating logical volume names, which was based on the same
functionality in blivet.