disk: EnsureRootFilesystem() helper function
Add a function that takes a partition table and adds a root filesystem,
but only if one does not already exist.  A root filesystem is any
mountable where the mountpoint is `/`.

The function decides whether to create a plain partition, a logical
volume, or a btrfs subvolume based on existing partitions and volumes in
the partition table.

This function is not currently used because our base partition tables
always have root filesystems, but it will be part of the new custom
partition table generator.