disk: EnsureBootPartition() helper function
Add a function that takes a partition table and adds a boot partition,
but only if one does not already exist.  A boot partition is any
mountable where the mountpoint is `/boot`.  Although the /boot
mountpoint must always be on a plain partition, this is not validated.

This function is not currently used because the NewPartitionTable()
function handles this automatically during layout conversions (in
ensureLVM() and ensureBtrfs()), but it will be part of the new custom
partition table generator.