github: rewrite integration workflow to post message
The new workflow doesn't fail if the integration test fails.  Instead it
posts a message notifying the author and reviewers of a PR that
compatibility/integration with osbuild-composer will fail when the
dependency is updated.

This message is only posted on a PR that causes the incompatibility
issue.  If compatibility is already failing on the merge base of the PR
(i.e. on main), the notice is not posted.

The comment in the workflow explains how things work in more detail.