container: add nested RHSM rpm test
This commit adds a test that ensures that access to subscribed
content works in a nested container similar to how we run the
nested container in bootc-image-builder. I.e. here we run in
a privilidged container that then runs a normal container to
depsolve against that.

The test works by running the already existing
`TestDNFJsonWorkWithSubscribedContent` inside the container to
double check that subscribed content is available both from the
host and when nesting.

The extra complication is that for efficiency we compile the
test binary on the host (to avoid having to have a go toolchain
inside the container) which means the container needs to have a
recent version of glibc (building our code with CGO_ENABLED=0
does not work) to support developing/running tests on f41.