disk: set the root and boot partition type correctly for dos
Define a new const string: DosLinuxTypeID, which is used for Linux
filesystems on dos/mbr partition tables.

In the new helper functions, use the partition table type to determine
which ID to use.  This means that the functions will return an error
when the partition table type is not set (or is set to an unknown
value).

Update the tests to match the new behaviour and to test the correct ID
selection.