disk: set the BIOS boot and ESP partition type correctly for dos
Define two new const strings: DosBIOSBootID and DosESPID, which are used
for the BIOS boot and the EFI system partitions respectively on dos/mbr
partition tables.

In the new helper functions, use the partition table type to determine
which ID to use.  This means that the functions will return an error
when the partition table type is not set (or is set to an unknown
value).