reporegistry: add new `LoadAllRepositoriesFromFS` and use in tests
This commit is a (hopefully) less controversial version of:
https://github.com/osbuild/images/pull/1037

Here only a new helper to load repository information from an
`fs.FS` is added and only used for loading the test repositories.

This will help to have a single test repository that can be
embedded accross osbuild-composer, image-builder and images.

It also removes the public "NewTestedDefault()" in favor of a
new `testrepos.New()` helper so that the `reporegistry` only
has non-test API left.