distro: add `OutputFilename` option to ImageOptions
This commit allows to change the filename when creating a
manifest from an image type. This is useful when doing a UI for users,
e.g. this will allow us to support:
```
$ image-builder build rhel-9 type:qcow2 --output foo.img
```
(this will also be useful for bootc-image-builder).

Well, osbuild itself will still put it under a directory when
doing main_cli.py:exports() but that is something orthogonal.